+++ vcfpy/reader.py
@@ -105,6 +105,12 @@ class Reader:
         if self.stream:
             self.stream.close()
 
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.close()
+
     def __iter__(self):
         return self