+++ vcfpy/writer.py
@@ -168,3 +168,9 @@ class Writer:
             return '.'
         else:
             return val
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.close()